Task3/requirements.txt
anyio==4.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
Automat==0.8.0
babel==2.16.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.4
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.8
cloud-init==24.4.1
cmake==3.31.1
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.6
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
document==1.0
drf-yasg==1.21.10
duckduckgo-search==7.2.1
entrypoints==0.3
et-xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.55.1
fsspec==2024.12.0
googlesearch-python==1.3.0
greenlet==0.4.15
h11==0.14.0
httpcore==1.0.6
httplib2==0.14.0
httpx==0.27.2
hyperlink==19.0.0
idna==2.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==16.10.1
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpatch==1.22
jsonpointer==2.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
keyring==18.0.1
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==5.3.1
macaroonbakery==1.3.1
markdown2==2.5.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==4.2.0
mpmath==1.3.0
msgpack==0.6.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netifaces==0.10.4
networkx==3.1
notebook==7.2.2
notebook-shim==0.2.4
notifications==0.3.2
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
primp==0.14.0
prometheus-client==0.21.0
prompt-toolkit==3.0.48
protobuf==3.6.1
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
pygments==2.18.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyOpenSSL==19.0.0
pyparsing==3.1.4
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pytest==8.3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-docx==1.1.2
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.2
PyYAML==5.3.1
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
selenium==4.26.1
Send2Trash==1.8.3
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
slugify==0.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sos==4.7.2
soupsieve==2.6
sqlparse==0.5.3
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.13.3
systemd-python==234
tenacity==9.0.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.2
torch==2.4.1
tornado==6.4.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
triton==3.0.0
Twisted==18.9.0
typing-extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.3
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
wsproto==1.2.0
xgboost==2.1.4
xmltodict==0.14.2
zipp==3.20.2
zope.interface==4.7.1
